@@ -1,78 +1,71 @@
 {
   "name": "ng",
-  "version": "3.0.1",
+  "version": "4.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
-    "start.ui.api": "ng serve --proxy-config proxy.api.conf.json",
-    "start.ui.mock": "ng serve --configuration=dev --proxy-config proxy.conf.js",
+    "start": "ng serve --proxy-config proxy.api.conf.json",
+    "start.remote": "ng serve --proxy-config proxy.api.conf.remote.json",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "dev:ssr": "ng run ng:serve-ssr",
-    "serve:ssr": "node dist/ng/server/main.js http://localhost:8080/",
-    "build:ssr": "ng build && ng run ng:server:production",
-    "prerender": "ng run ng:prerender"
+    "serve:ssr:ng": "node dist/ng/server/server.mjs http://localhost:8080/web4"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.0.0",
-    "@angular/cdk": "^16.1.2",
-    "@angular/common": "^16.0.0",
-    "@angular/compiler": "^16.0.0",
-    "@angular/core": "^16.0.0",
-    "@angular/flex-layout": "^14.0.0-beta.40",
-    "@angular/forms": "^16.0.0",
-    "@angular/material": "^14.0.5",
-    "@angular/material-moment-adapter": "^14.0.5",
-    "@angular/platform-browser": "^16.0.0",
-    "@angular/platform-browser-dynamic": "^16.0.0",
-    "@angular/platform-server": "^16.0.0",
-    "@angular/router": "^16.0.0",
-    "@asymmetrik/ngx-leaflet": "^16.0.1",
-    "@asymmetrik/ngx-leaflet-markercluster": "^16.0.0",
-    "@ngu/carousel": "^7.1.3",
-    "@nguniversal/express-engine": "^16.1.0",
-    "@ngx-translate/core": "^15.0.0",
-    "@ngx-translate/http-loader": "^8.0.0",
-    "echarts": "^5.5.1",
-    "express": "^4.21.1",
-    "hammerjs": "^2.0.8",
-    "i": "^0.3.7",
+    "@angular/animations": "^20.0.5",
+    "@angular/cdk": "^20.0.4",
+    "@angular/common": "^20.0.5",
+    "@angular/compiler": "^20.0.5",
+    "@angular/core": "^20.0.5",
+    "@angular/forms": "^20.0.5",
+    "@angular/material": "^20.0.4",
+    "@angular/material-moment-adapter": "^20.1.5",
+    "@angular/platform-browser": "^20.0.5",
+    "@angular/platform-server": "^20.0.5",
+    "@angular/router": "^20.0.5",
+    "@angular/ssr": "^20.0.4",
+    "@bluehalo/ngx-leaflet": "^20.0.0",
+    "@bluehalo/ngx-leaflet-markercluster": "^20.0.3",
+    "@ngu/carousel": "^20.0.1",
+    "@ngx-translate/core": "^17.0.0",
+    "@ngx-translate/http-loader": "^17.0.0",
+    "@types/leaflet.markercluster": "^1.5.5",
+    "angular-material-css-vars": "^9.1.1",
+    "axios": "^1.11.0",
+    "echarts": "^5.6.0",
+    "express": "^5.1.0",
+    "heatmap.js": "^2.0.5",
     "leaflet": "^1.9.4",
     "leaflet-heatmap": "^1.0.0",
-    "leaflet.fullscreen": "^2.4.0",
+    "leaflet.fullscreen": "^4.0.0",
     "leaflet.locationfilter": "^0.1.0",
     "leaflet.markercluster": "^1.5.3",
     "leaflet.polylinemeasure": "^3.0.0",
-    "moment": "^2.29.4",
-    "ng-recaptcha": "^12.0.1",
-    "ngx-echarts": "^16.2.0",
-    "npm": "^10.9.2",
-    "request": "^2.88.2",
+    "moment": "^2.30.1",
+    "ng-recaptcha-2": "^16.0.1",
+    "ngx-echarts": "^20.0.1",
+    "ngx-flexible-layout": "^20.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "wicket": "^1.3.8",
-    "zone.js": "~0.13.0"
+    "wicket": "^1.3.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.16",
-    "@angular/cli": "~16.0.4",
-    "@angular/compiler-cli": "^16.0.0",
-    "@nguniversal/builders": "^16.1.0",
-    "@types/express": "^4.17.0",
-    "@types/hammerjs": "^2.0.41",
-    "@types/jasmine": "~4.3.0",
-    "@types/leaflet": "^1.9.3",
-    "@types/leaflet.markercluster": "^1.5.1",
-    "@types/node": "^14.18.63",
-    "jasmine-core": "~4.6.0",
+    "@angular/build": "^20.0.4",
+    "@angular/cli": "^20.0.4",
+    "@angular/compiler-cli": "^20.0.5",
+    "@types/express": "^5.0.1",
+    "@types/jasmine": "~5.1.0",
+    "@types/leaflet": "^1.9.19",
+    "@types/leaflet.markercluster": "^1.5.5",
+    "@types/leaflet.polylinemeasure": "^1.0.4",
+    "@types/node": "^20.17.19",
+    "@types/wicket": "^1.3.0",
+    "jasmine-core": "~5.7.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "ts-node": "^10.9.1",
-    "typescript": "~5.0.2"
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.8.2"
   }
 }
