@@ -1,6 +1,7 @@
-import { AppService } from 'src/app/app.service';
-import { Component, OnInit, Input } from '@angular/core';
+
+import { Component, OnInit, effect, input, signal } from '@angular/core';
 import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
+import { AppService } from '../../app.service';
 
 @Component({
   selector: 'app-free-text',
@@ -9,24 +10,31 @@ import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
 })
 export class FreeTextComponent implements OnInit {
 
-  @Input() id: string;
-  text: SafeHtml;
+  readonly id = input<string>();
+  text = signal<SafeHtml>('');
 
   constructor(
     private sanitized: DomSanitizer,
     private service: AppService
-  ) { }
+  ) { 
+    effect(() => {
+      const id = this.id();
+      if (id) {
+        this.setText();
+      }
+    })
+  }
 
   ngOnInit(): void {
     this.service.currentLang.subscribe(r => {
       this.setText();
     });
-    this.setText();
+    
   }
 
   setText() {
-    this.service.getText(this.id).subscribe(t => {
-      this.text = this.sanitized.bypassSecurityTrustHtml(t);
+    this.service.getText(this.id()).subscribe(t => {
+      this.text.set(this.sanitized.bypassSecurityTrustHtml(t));
     });
   }
 
