@@ -1,21 +1,35 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { AppService } from 'src/app/app.service';
-import { AppConfiguration } from 'src/app/app-configuration';
+import { CommonModule } from '@angular/common';
+import { Component, OnInit, computed, input } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { MatIconModule } from '@angular/material/icon';
 import { Router } from '@angular/router';
+import { TranslateModule } from '@ngx-translate/core';
+import { FlexLayoutModule } from 'ngx-flexible-layout';
+import { AppConfiguration } from '../../app-configuration';
+import { AppService } from '../../app.service';
+import { MatButtonModule } from '@angular/material/button';
 
 @Component({
+  imports: [
+    TranslateModule, CommonModule, FormsModule, FlexLayoutModule,
+    MatIconModule, MatButtonModule
+  ],
   selector: 'app-inline-filter',
   templateUrl: './inline-filter.component.html',
-  styleUrls: ['./inline-filter.component.scss']
+  styleUrls: ['./inline-filter.component.scss'],
+  preserveWhitespaces: false
 })
 export class InlineFilterComponent implements OnInit {
 
-  @Input() field: string;
-  @Input() value: any;
-  @Input() heslar: string;
-  @Input() isChild: boolean;
-  @Input() isDate = false;
-  @Input() isYear= false;
+  readonly field = input<string>();
+  readonly value = input<any>();
+  public strVal = computed<string>(() => 
+    (this.value()?.id ? this.value().id : this.value()) + ''
+  );
+  readonly heslar = input<string>();
+  readonly isChild = input<boolean>();
+  readonly isDate = input(false);
+  readonly isYear = input(false);
   isDocument: boolean;
 
   constructor(
@@ -28,13 +42,15 @@ export class InlineFilterComponent implements OnInit {
   }
 
   addFilter(event: any) {
-    let v = this.value + '';
-    const filter = this.config.filterFields.find(ff => ff.field === this.field);
+    const val = this.value();
+    console.log(this.field(), val, val.id)
+    let v = (val.id ? val.id : val) + '';
+    const filter = this.config.filterFields.find(ff => ff.field === this.field());
     if (filter && filter.type === 'date') {
     // if (this.config.dateFacets.includes(this.field)) {
       v += ',' + v;
     }
-    this.service.addFilter(this.field, v, 'and');
+    this.service.addFilter(this.field(), v, 'and');
     event.stopPropagation();
   }
 
