@@ -1,27 +1,42 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
-import { AppConfiguration } from 'src/app/app-configuration';
-import { AppService } from 'src/app/app.service';
-import { AppState } from 'src/app/app.state';
+import { CommonModule } from '@angular/common';
+import { Component, EventEmitter, Output, input, signal } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { MatIconModule } from '@angular/material/icon';
+import { Router, RouterModule } from '@angular/router';
+import { TranslateModule } from '@ngx-translate/core';
+import { FlexLayoutModule } from 'ngx-flexible-layout';
+import { AppConfiguration } from '../../app-configuration';
+import { AppService } from '../../app.service';
+import { AppState } from '../../app.state';
+import { MatButtonModule } from '@angular/material/button';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { BibtextDialogComponent } from '../bibtext-dialog/bibtext-dialog.component';
 import { DocumentDialogComponent } from '../document-dialog/document-dialog.component';
 import { FeedbackDialogComponent } from '../feedback-dialog/feedback-dialog.component';
-import { Router } from '@angular/router';
-import { BibtextDialogComponent } from '../bibtext-dialog/bibtext-dialog.component';
 
 @Component({
+  imports: [
+    TranslateModule, CommonModule, FormsModule, FlexLayoutModule,
+    RouterModule, MatDialogModule, MatButtonModule,
+    MatIconModule, MatButtonModule, MatMenuModule, MatTooltipModule
+  ],
   selector: 'app-result-actions',
   templateUrl: './result-actions.component.html',
   styleUrls: ['./result-actions.component.scss']
 })
 export class ResultActionsComponent {
 
-  @Input() result: any;
-  @Input() bibTex: string;
-  @Input() isDocumentDialogOpen: boolean;
-  @Input() detailExpanded: boolean;
-  @Input() inDocument: boolean;
-  @Input() mapDetail: boolean;
-  @Input() ident_cely_api: any;
+  readonly result = input<any>();
+  readonly bibTex = input<string>();
+  readonly isDocumentDialogOpen = input<boolean>();
+  readonly detailExpanded = input<boolean>();
+  readonly inDocument = input<boolean>();
+  readonly mapDetail = input<boolean>();
+  readonly ident_cely_api = input<any>();
+
+  isFav = signal<boolean>(false);
 
   @Output() onToggleDetail  = new EventEmitter<string>();
 
@@ -35,8 +50,14 @@ export class ResultActionsComponent {
     public config: AppConfiguration
   ) { }
 
+  ngOnInit() {
+    this.isFav.set(this.result().isFav)
+  }
+
   apiIdentCely(item:{label: string, metadataPrefix: string, url: string, useParent: boolean}) {
-    return (item.useParent && this.ident_cely_api) ? this.ident_cely_api : this.result.ident_cely;
+    const ident_cely_api = this.ident_cely_api();
+    // console.log(item.useParent)
+    return (item.useParent && ident_cely_api) ? ident_cely_api : this.result().ident_cely;
   }
 
   toggleDetail() {
@@ -44,13 +65,14 @@ export class ResultActionsComponent {
   }
 
   toggleFav() {
-    if (this.result.isFav) {
-      this.service.removeFav(this.result.ident_cely).subscribe(res => {
-        this.result.isFav = false;
+    const result = this.result();
+    if (this.isFav()) {
+      this.service.removeFav(result.ident_cely).subscribe(res => {
+        this.isFav.set(false);
       });
     } else {
-      this.service.addFav(this.result.ident_cely).subscribe(res => {
-        this.result.isFav = true;
+      this.service.addFav(result.ident_cely).subscribe(res => {
+        this.isFav.set(true);
       });
     }
   }
@@ -58,15 +80,15 @@ export class ResultActionsComponent {
   openDocument() {
     this.state.dialogRef = this.dialog.open(DocumentDialogComponent, {
       width: '900px',
-      data: this.result,
-      panelClass: 'app-document-dialog'
+      data: this.result(),
+      panelClass: 'app-card-result-dialog'
     });
   }
 
   openFeedback() {
     this.state.dialogRef = this.dialog.open(FeedbackDialogComponent, {
       width: '900px',
-      data: this.result.ident_cely,
+      data: this.result().ident_cely,
       panelClass: 'app-feedback-dialog'
     });
   }
@@ -74,7 +96,7 @@ export class ResultActionsComponent {
   showCitation() {
     this.state.dialogRef = this.dialog.open(BibtextDialogComponent, {
       width: '900px',
-      data: {result: this.result, link: this.config.serverUrl + 'id/' + this.result.ident_cely},
+      data: {result: this.result(), link: this.config.serverUrl + 'id/' + this.result().ident_cely},
       panelClass: 'app-feedback-dialog'
     });
     
