@@ -1,8 +1,22 @@
-import { Component, OnInit, Inject } from '@angular/core';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { AppService } from 'src/app/app.service';
+
+import { Component, OnInit, Inject, forwardRef } from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
+import { MatDialogRef, MAT_DIALOG_DATA, MatDialogModule } from '@angular/material/dialog';
+import { MatIconModule } from '@angular/material/icon';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { TranslateModule } from '@ngx-translate/core';
+import { AppService } from '../../app.service';
+import { EntityContainer } from '../../entities/entity-container/entity-container';
 
 @Component({
+  imports: [
+    TranslateModule,
+    MatDialogModule,
+    MatIconModule,
+    MatButtonModule,
+    MatTooltipModule,
+    forwardRef(() => EntityContainer),
+],
   selector: 'app-document-dialog',
   templateUrl: './document-dialog.component.html',
   styleUrls: ['./document-dialog.component.scss']
@@ -25,6 +39,10 @@ export class DocumentDialogComponent implements OnInit {
     this.service.getId(this.result.ident_cely).subscribe((res: any) => {
       this.result = res.response.docs[0];
     });
+  }
+
+  isResultEmpty(obj: any) {
+    return (!obj || (Object.keys(obj).length === 0));
   }
 
 }
