@@ -1,12 +1,26 @@
 import { Component, OnInit } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { FormControl, FormsModule } from '@angular/forms';
 import { Subscription } from 'rxjs/internal/Subscription';
 import { Subject } from 'rxjs';
 import { debounceTime, distinctUntilChanged } from 'rxjs/operators';
-import { AppState } from 'src/app/app.state';
-import { AppService } from 'src/app/app.service';
+import { MatCardModule } from '@angular/material/card';
+import { MatIconModule } from '@angular/material/icon';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { RouterModule } from '@angular/router';
+import { TranslateModule } from '@ngx-translate/core';
+import { FlexLayoutModule } from 'ngx-flexible-layout';
+import { AppService } from '../../../app.service';
+import { AppState } from '../../../app.state';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatButtonModule } from '@angular/material/button';
+import { MatInputModule } from '@angular/material/input';
 
 @Component({
+  imports: [
+    TranslateModule, RouterModule, FlexLayoutModule, FormsModule, MatInputModule,
+    MatCardModule, MatIconModule, MatTooltipModule, MatFormFieldModule,
+    MatButtonModule
+],
   selector: 'app-facets-search',
   templateUrl: './facets-search.component.html',
   styleUrls: ['./facets-search.component.scss']
@@ -36,7 +50,7 @@ export class FacetsSearchComponent implements OnInit {
       });
   }
 
-  changed(e) {
+  changed(e: any) {
     this.modelChanged.next(e.target.value);
 }
 
