@@ -1,12 +1,24 @@
-import { Crumb } from 'src/app/shared/crumb';
-import { AppService } from 'src/app/app.service';
-import { Router, ActivatedRoute } from '@angular/router';
+
+import { Router, ActivatedRoute, RouterModule } from '@angular/router';
 import { Component, OnInit } from '@angular/core';
-import { AppConfiguration } from 'src/app/app-configuration';
-import { AppState } from 'src/app/app.state';
-import { Sort } from 'src/app/shared/config';
+import { MatCardModule } from '@angular/material/card';
+import { MatIconModule } from '@angular/material/icon';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { TranslateModule } from '@ngx-translate/core';
+import { FlexLayoutModule } from 'ngx-flexible-layout';
+import { AppConfiguration } from '../../../app-configuration';
+import { AppService } from '../../../app.service';
+import { AppState } from '../../../app.state';
+import { Crumb } from '../../../shared/crumb';
+import { MatListModule } from '@angular/material/list';
+import { MatButtonModule } from '@angular/material/button';
 
 @Component({
+  imports: [
+    TranslateModule, RouterModule, FlexLayoutModule, 
+    MatCardModule, MatIconModule, MatTooltipModule, MatListModule,
+    MatButtonModule
+],
   selector: 'app-facets-used',
   templateUrl: './facets-used.component.html',
   styleUrls: ['./facets-used.component.scss']
@@ -22,9 +34,17 @@ export class FacetsUsedComponent implements OnInit {
 
   ngOnInit(): void {
     
+      this.service.setCrumbs(this.route.snapshot.queryParamMap);
     this.service.currentLang.subscribe(() => {
       this.service.setCrumbs(this.route.snapshot.queryParamMap);
     });
+  }
+
+  removeInfavorites() {
+    
+    const params: any = {};
+    params['inFavorites'] = null;
+    this.router.navigate([], { queryParams: params, queryParamsHandling: 'merge' });
   }
 
   removeCommonFacet(name: string) {
@@ -58,7 +78,13 @@ export class FacetsUsedComponent implements OnInit {
     this.state.breadcrumbs.forEach((c: Crumb) => {
       q[c.field] = null;
     });
+
+    this.config.commonFacets.forEach(cf => {
+      q[cf.name] = null;
+    });
     q.page = 0;
+    q['inFavorites'] = null;
+    this.state.breadcrumbs = [];
     this.router.navigate([], { queryParams: q, queryParamsHandling: 'merge' });
   }
 }
