@@ -1,16 +1,6 @@
-import { enableProdMode } from '@angular/core';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { bootstrapApplication } from '@angular/platform-browser';
+import { appConfig } from './app/app.config';
+import { App } from './app/app';
 
-import { AppBrowserModule } from './app/app.browser.module';
-// import { AppModule } from './app/app.module';
-import { environment } from './environments/environment';
-
-if (environment.production) {
-  enableProdMode();
-}
-
-document.addEventListener('DOMContentLoaded', () => {
-  // platformBrowserDynamic().bootstrapModule(AppModule)
-  platformBrowserDynamic().bootstrapModule(AppBrowserModule)
-  .catch(err => console.error(err));
-});
+bootstrapApplication(App, appConfig)
+  .catch((err) => console.error(err));
